+++ src/components/Header/style.tsx
@@ -17,3 +17,11 @@ export const themeIcon = {
 
 export const darkLogo = 'https://gw.alipayobjects.com/zos/bmw-prod/9ecb2822-1592-4cb0-a087-ce0097fef2ca.svg'
 export const lightLogo = 'https://gw.alipayobjects.com/zos/bmw-prod/e146116d-c65a-4306-a3d2-bb8d05e1c49b.svg'
+
+export const civitaiLogo = (
+  <span role="img" className="anticon civitai">
+    <svg viewBox="0 0 16 16" width="1em" height="1em" fill="currentColor">
+      <path d="M2 4.5L8 1l6 3.5v7L8 15l-6-3.5v-7zm6-1.194L3.976 5.653v4.694L8 12.694l4.024-2.347V5.653L8 3.306zm0 1.589l2.662 1.552v.824H9.25L8 6.54l-1.25.73v1.458L8 9.46l1.25-.73h1.412v.824L8 11.105 5.338 9.553V6.447L8 4.895z" />
+    </svg>
+  </span>
+)