+++ src/components/Sidebar/Prompt.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useEffect, useState } from 'react'
+import React, { useCallback, useState } from 'react'
 import styled from 'styled-components'
 import TagList, { PromptType, TagItem } from './TagList'
 import { formatPrompt } from './utils'
@@ -9,7 +9,12 @@ const View = styled.div`
   gap: 8px;
 `
 
-const Reload = styled.button`
+const BtnGroup = styled.div`
+  display: flex;
+  gap: 8px;
+`
+
+const Btn = styled.button`
   display: flex;
   align-items: center;
   justify-content: center;
@@ -22,6 +27,7 @@ const Reload = styled.button`
   line-height: var(--line-sm);
   padding: var(--input-padding);
   cursor: pointer;
+  flex: 1;
 `
 
 interface PromptProps {
@@ -33,30 +39,31 @@ const Prompt: React.FC<PromptProps> = ({ type }) => {
 
   const id = type === 'positive' ? "[id$='2img_prompt'] textarea" : "[id$='2img_neg_prompt'] textarea"
 
-  useEffect(() => {
-    onUiUpdate(() => {
+  const getValue = useCallback(() => {
+    try {
       const textarea: HTMLTextAreaElement | any = get_uiCurrentTabContent().querySelector(id)
-      if (textarea) textarea.addEventListener('change', (e: any) => setTags(formatPrompt(e.target.value)), false)
-    })
+      if (textarea) setTags(formatPrompt(textarea.value))
+    } catch {}
   }, [])
 
-  useEffect(() => {
-    if (!tags) return
+  const setValue = useCallback(() => {
     try {
       const textarea: HTMLTextAreaElement | any = get_uiCurrentTabContent().querySelector(id)
       if (textarea) textarea.value = tags.map((t) => t.text).join(', ')
     } catch {}
   }, [tags])
 
-  const getValue = useCallback(() => {
-    const textarea: HTMLTextAreaElement | any = get_uiCurrentTabContent().querySelector(id)
-    if (textarea) setTags(formatPrompt(textarea.value))
-  }, [])
-
   return (
     <View>
       <TagList type={type} tags={tags} setTags={setTags} />
-      <Reload onClick={getValue}>🔄</Reload>
+      <BtnGroup>
+        <Btn title="Load Prompt" onClick={getValue}>
+          🔄
+        </Btn>
+        <Btn title="Set Prompt" onClick={setValue}>
+          ➡️
+        </Btn>
+      </BtnGroup>
     </View>
   )
 }